Fix typo to sftp_to_wasb file name Adds the information from extra package to CONTRIBUTING.rst
