+++ tests/providers/microsoft/azure/transfers/test_sftp_to_wasb.py
@@ -17,17 +17,15 @@
 # under the License.
 #
 
-
-import unittest
 import os
+import unittest
 
 import mock
 
 from airflow import AirflowException
-from airflow.providers.microsoft.azure.transfers.stfp_to_wasb import STFPToWasbOperator
-from airflow.providers.microsoft.azure.transfers.stfp_to_wasb import SFTP_FILE_PATH
-from airflow.providers.microsoft.azure.transfers.stfp_to_wasb import BLOB_NAME
-
+from airflow.providers.microsoft.azure.transfers.sftp_to_wasb import SFTPToWasbOperator, \
+    SFTP_FILE_PATH, \
+    BLOB_NAME
 
 TASK_ID = "test-gcs-to-sftp-operator"
 WASB_CONN_ID = "wasb_default"
@@ -46,7 +44,7 @@ EXPECTED_BLOB_NAME = "test_object3.json"
 class TestSFTPToWasbOperator(unittest.TestCase):
 
     def test_init(self):
-        operator = STFPToWasbOperator(
+        operator = SFTPToWasbOperator(
             task_id=TASK_ID,
             sftp_source_path=SOURCE_OBJECT_NO_WILDCARD,
             sftp_conn_id=SFTP_CONN_ID,
@@ -60,10 +58,10 @@ class TestSFTPToWasbOperator(unittest.TestCase):
         self.assertEqual(operator.container_name, CONTAINER_NAME)
         self.assertEqual(operator.wasb_conn_id, WASB_CONN_ID)
 
-    @mock.patch('airflow.providers.microsoft.azure.transfers.stfp_to_wasb.WasbHook',
+    @mock.patch('airflow.providers.microsoft.azure.transfers.sftp_to_wasb.WasbHook',
                 autospec=True)
     def test_execute_more_than_one_wildcard_exception(self, mock_hook):
-        operator = STFPToWasbOperator(
+        operator = SFTPToWasbOperator(
             task_id=TASK_ID,
             sftp_source_path=SOURCE_OBJECT_MULTIPLE_WILDCARDS,
             sftp_conn_id=SFTP_CONN_ID,
@@ -78,15 +76,15 @@ class TestSFTPToWasbOperator(unittest.TestCase):
         err = cm.exception
         self.assertIn("Only one wildcard '*' is allowed", str(err))
 
-    @mock.patch('airflow.providers.microsoft.azure.transfers.stfp_to_wasb.WasbHook')
-    @mock.patch('airflow.providers.microsoft.azure.transfers.stfp_to_wasb.SFTPHook')
+    @mock.patch('airflow.providers.microsoft.azure.transfers.sftp_to_wasb.WasbHook')
+    @mock.patch('airflow.providers.microsoft.azure.transfers.sftp_to_wasb.SFTPHook')
     def test_get_sftp_files_map_with_wildcard(self, sftp_hook, mock_hook):
         sftp_hook.return_value.get_tree_map.return_value = [
             [SOURCE_OBJECT_NO_WILDCARD, SOURCE_OBJECT_NO_WILDCARD],
             [],
             [],
         ]
-        operator = STFPToWasbOperator(
+        operator = SFTPToWasbOperator(
             task_id=TASK_ID,
             sftp_source_path=WILDCARD_PATH,
             sftp_conn_id=SFTP_CONN_ID,
@@ -102,15 +100,15 @@ class TestSFTPToWasbOperator(unittest.TestCase):
         self.assertEqual(len(files), 2, "not matched at expected found files")
         self.assertEqual(files[0][BLOB_NAME], EXPECTED_BLOB_NAME, "expected blob name not matched")
 
-    @mock.patch('airflow.providers.microsoft.azure.transfers.stfp_to_wasb.WasbHook')
-    @mock.patch('airflow.providers.microsoft.azure.transfers.stfp_to_wasb.SFTPHook')
+    @mock.patch('airflow.providers.microsoft.azure.transfers.sftp_to_wasb.WasbHook')
+    @mock.patch('airflow.providers.microsoft.azure.transfers.sftp_to_wasb.SFTPHook')
     def test_get_sftp_files_map_no_wildcard(self, sftp_hook, mock_hook):
         sftp_hook.return_value.get_tree_map.return_value = [
             [SOURCE_OBJECT_NO_WILDCARD],
             [],
             [],
         ]
-        operator = STFPToWasbOperator(
+        operator = SFTPToWasbOperator(
             task_id=TASK_ID,
             sftp_source_path=SOURCE_OBJECT_NO_WILDCARD,
             sftp_conn_id=SFTP_CONN_ID,
@@ -126,10 +124,10 @@ class TestSFTPToWasbOperator(unittest.TestCase):
         self.assertEqual(len(files), 1, "no matched at expected found files")
         self.assertEqual(files[0][BLOB_NAME], BLOB_NAME, "expected blob name not matched")
 
-    @mock.patch('airflow.providers.microsoft.azure.transfers.stfp_to_wasb.WasbHook')
-    @mock.patch('airflow.providers.microsoft.azure.transfers.stfp_to_wasb.SFTPHook')
+    @mock.patch('airflow.providers.microsoft.azure.transfers.sftp_to_wasb.WasbHook')
+    @mock.patch('airflow.providers.microsoft.azure.transfers.sftp_to_wasb.SFTPHook')
     def test_upload_wasb(self, sftp_hook, mock_hook):
-        operator = STFPToWasbOperator(
+        operator = SFTPToWasbOperator(
             task_id=TASK_ID,
             sftp_source_path=SOURCE_OBJECT_NO_WILDCARD,
             sftp_conn_id=SFTP_CONN_ID,
@@ -154,11 +152,11 @@ class TestSFTPToWasbOperator(unittest.TestCase):
 
         self.assertEqual(len(files), 1, "no matched at expected uploaded files")
 
-    @mock.patch('airflow.providers.microsoft.azure.transfers.stfp_to_wasb.SFTPHook')
+    @mock.patch('airflow.providers.microsoft.azure.transfers.sftp_to_wasb.SFTPHook')
     def test_sftp_move(self, sftp_hook):
         sftp_mock = sftp_hook.return_value
 
-        operator = STFPToWasbOperator(
+        operator = SFTPToWasbOperator(
             task_id=TASK_ID,
             sftp_source_path=SOURCE_OBJECT_NO_WILDCARD,
             sftp_conn_id=SFTP_CONN_ID,
@@ -177,11 +175,11 @@ class TestSFTPToWasbOperator(unittest.TestCase):
             ]
         )
 
-    @mock.patch('airflow.providers.microsoft.azure.transfers.stfp_to_wasb.SFTPHook')
+    @mock.patch('airflow.providers.microsoft.azure.transfers.sftp_to_wasb.SFTPHook')
     def test_no_sftp_move(self, sftp_hook):
         sftp_mock = sftp_hook.return_value
 
-        operator = STFPToWasbOperator(
+        operator = SFTPToWasbOperator(
             task_id=TASK_ID,
             sftp_source_path=SOURCE_OBJECT_NO_WILDCARD,
             sftp_conn_id=SFTP_CONN_ID,
@@ -196,8 +194,8 @@ class TestSFTPToWasbOperator(unittest.TestCase):
 
         sftp_mock.delete_file.assert_not_called()
 
-    @mock.patch('airflow.providers.microsoft.azure.transfers.stfp_to_wasb.WasbHook')
-    @mock.patch('airflow.providers.microsoft.azure.transfers.stfp_to_wasb.SFTPHook')
+    @mock.patch('airflow.providers.microsoft.azure.transfers.sftp_to_wasb.WasbHook')
+    @mock.patch('airflow.providers.microsoft.azure.transfers.sftp_to_wasb.SFTPHook')
     def test_execute(self, sftp_hook, mock_hook):
         sftp_hook.return_value.get_tree_map.return_value = [
             [SOURCE_OBJECT_NO_WILDCARD],
@@ -205,7 +203,7 @@ class TestSFTPToWasbOperator(unittest.TestCase):
             [],
         ]
 
-        operator = STFPToWasbOperator(
+        operator = SFTPToWasbOperator(
             task_id=TASK_ID,
             sftp_source_path=WILDCARD_FILE_NAME,
             sftp_conn_id=SFTP_CONN_ID,
